refactor(workbench-cli): move the undeploy API wrappers into services
